api/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
python-dotenv
requests
psycopg2-binary
Faker
